(feat) rename components
